requirements.txt
pytorch-lightning==0.7.1
PyYAML==5.1.2
tensorboard==2.1.0
tensorboardX==1.6
torch==1.2.0
numpy
bleu
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.4.0